requirements.txt
streamlit
streamlit_chat
google.generativeai